+++ LingoClash/LingoClash/ProfileView/ProfileViewModel.swift
@@ -37,9 +37,6 @@ final class ProfileViewModel {
         }
     }
     
-    var firstName: String?
-    var lastName: String?
-    
     func refreshProfile() {
         // TODO: get user profile
         self.firstName = "John"
